add validation

+++ src/main/java/com/innowise/accountingsystem/service/EmployeeService.java
@@ -9,6 +9,7 @@ import com.innowise.accountingsystem.entity.Employee;
 import com.innowise.accountingsystem.entity.Role;
 import com.innowise.accountingsystem.mapper.EmployeeMapper;
 import com.innowise.accountingsystem.util.PasswordEncoder;
+import com.innowise.accountingsystem.validation.EmployeeValidator;
 import org.mapstruct.factory.Mappers;
 
 import java.util.List;
@@ -21,11 +22,13 @@ public class EmployeeService {
     private final EmployeeDao employeeDao;
     private final EmployeeMapper employeeMapper;
     private final PasswordEncoder passwordEncoder;
+    private final EmployeeValidator employeeValidator;
 
     private EmployeeService() {
         this.employeeDao = EmployeeDao.getInstance();
         this.employeeMapper = Mappers.getMapper(EmployeeMapper.class);
         this.passwordEncoder = PasswordEncoder.getInstance();
+        this.employeeValidator = EmployeeValidator.getInstance();
     }
 
     public static EmployeeService getInstance() {
@@ -33,36 +36,41 @@ public class EmployeeService {
     }
 
     public Optional<EmployeeDto> addEmployee(CreateEmployeeDto createEmployeeDto) {
-        Optional<EmployeeDto> optionalEmployee;
-        //validate
-        Employee employee = employeeMapper.fromCreateEmployeeDto(createEmployeeDto);
-        employee.setPassword(passwordEncoder.encode(employee.getPassword()));
-        employee.setRole(Role.USER);
+        Optional<EmployeeDto> optionalEmployee = Optional.empty();
 
-        try {
-            employeeDao.save(employee);
-            EmployeeDto employeeDto = employeeMapper.toDto(employee);
-            optionalEmployee = Optional.of(employeeDto);
+        if (employeeValidator.isCreateEmployeeDtoValid(createEmployeeDto)) {
+            Employee employee = employeeMapper.fromCreateEmployeeDto(createEmployeeDto);
+            employee.setPassword(passwordEncoder.encode(employee.getPassword()));
+            employee.setRole(Role.USER);
 
-        } catch (DaoException e) {
-            //log
-            throw new ServiceException(e);
+            try {
+                employeeDao.save(employee);
+                EmployeeDto employeeDto = employeeMapper.toDto(employee);
+                optionalEmployee = Optional.of(employeeDto);
+            } catch (DaoException e) {
+                //log
+                throw new ServiceException(e);
+            }
         }
 
         return optionalEmployee;
     }
 
     public Optional<EmployeeDto> getEmployeeById(String id) {
-        Optional<EmployeeDto> optionalEmployee;
-        //validate id
-        try {
-            Optional<Employee> employee = employeeDao.findById(Long.parseLong(id));
-            optionalEmployee = employee.map(employeeMapper::toDto);
-            return optionalEmployee;
-        } catch (DaoException e) {
-            //log.error
-            throw new ServiceException(e);
+        Optional<EmployeeDto> optionalEmployee = Optional.empty();
+
+        if (employeeValidator.isIdValid(id)) {
+            try {
+                Optional<Employee> employee = employeeDao.findById(Long.parseLong(id));
+                optionalEmployee = employee.map(employeeMapper::toDto);
+
+            } catch (DaoException e) {
+                //log.error
+                throw new ServiceException(e);
+            }
         }
+
+        return optionalEmployee;
     }
 
     public List<EmployeeDto> getAllEmployees() {
@@ -75,50 +83,59 @@ public class EmployeeService {
     public Optional<EmployeeDto> getEmployeeByEmailAndPassword(String email, String password) {
         Optional<EmployeeDto> optionalEmployee = Optional.empty();
 
-        try {
-            Optional<Employee> employee = employeeDao.findByEmail(email);
-            if (employee.isPresent()) {
-                String hash = employee.get().getPassword();
-                if (passwordEncoder.verify(hash, password)) {
-                    optionalEmployee = employee.map(employeeMapper::toDto);
+        if (employeeValidator.isEmailValid(email) && employeeValidator.isPasswordValid(password)) {
+            try {
+                Optional<Employee> employee = employeeDao.findByEmail(email);
+                if (employee.isPresent()) {
+                    String hash = employee.get().getPassword();
+                    if (passwordEncoder.verify(hash, password)) {
+                        optionalEmployee = employee.map(employeeMapper::toDto);
+                    }
                 }
-            }
 
-            return optionalEmployee;
-        } catch (DaoException e) {
-            //log.error
-            throw new ServiceException(e);
+
+            } catch (DaoException e) {
+                //log.error
+                throw new ServiceException(e);
+            }
         }
+
+        return optionalEmployee;
     }
 
     public Optional<EmployeeDto> updateEmployee(EmployeeDto employeeDto) {
         Optional<EmployeeDto> optionalEmployee = Optional.empty();
-        Employee employee = employeeMapper.toEntity(employeeDto);
 
-        try {
-            Optional<Employee> updatedEmployee = employeeDao.update(employee);
+        if (employeeValidator.isEmployeeDtoValid(employeeDto)) {
+            Employee employee = employeeMapper.toEntity(employeeDto);
 
-            if (updatedEmployee.isPresent()) {
-                EmployeeDto updatedEmployeeDto = employeeMapper.toDto(employee);
-                optionalEmployee = Optional.of(updatedEmployeeDto);
+            try {
+                Optional<Employee> updatedEmployee = employeeDao.update(employee);
+
+                if (updatedEmployee.isPresent()) {
+                    EmployeeDto updatedEmployeeDto = employeeMapper.toDto(employee);
+                    optionalEmployee = Optional.of(updatedEmployeeDto);
+                }
+            } catch (DaoException e) {
+                //log.error
+                throw new ServiceException(e);
             }
-        } catch (DaoException e) {
-            //log.error
-            throw new ServiceException(e);
         }
 
         return optionalEmployee;
     }
 
     public boolean deleteEmployee(String id) {
-        //validate id
-
-        try {
-            Optional<Employee> employee = employeeDao.findById(Long.parseLong(id));
-            return employeeDao.delete(Long.parseLong(id));
-        } catch (DaoException e) {
-            //log.error
-            throw new ServiceException(e);
+        if (employeeValidator.isIdValid(id)) {
+            try {
+                employeeDao.findById(Long.parseLong(id));
+                return employeeDao.delete(Long.parseLong(id));
+            } catch (DaoException e) {
+                //log.error
+                throw new ServiceException(e);
+            }
+        } else {
+            return false;
         }
     }
 }
